backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-telegram-bot[rate-limiter]==20.6
sqlalchemy==2.0.21
python-dotenv==1.0.0
openai==0.28.1
pydantic==1.10.12
aiohttp==3.8.5; platform_system!="Darwin"
aiohttp==3.7.4; platform_system=="Darwin"
python-multipart==0.0.6
typing-extensions>=4.7.1
pyrogram==2.0.106
tgcrypto==1.2.5